+++ packages/plugin-swr/src/generators/mutationGenerator.tsx
@@ -165,6 +165,7 @@ export const mutationGenerator = createReactGenerator<PluginSwr>({
             pathParamsType={options.pathParamsType}
             mutationKeyName={mutationKey.name}
             mutationKeyTypeName={mutationKey.typeName}
+            paramsToTrigger={options.mutation.paramsToTrigger}
           />
         )}
       </File>